application/requirements.txt
pdfplumber
Pillow
Pyrebase4
python-dotenv
reportlab
requests
streamlit
streamlit_extras
web3
cryptography
